Update no-console rule for Ember
Ember is getting rid of the logger, which really was just a thing wrapper around the console and existed to patch up cross-browser inconsistencies. So, modify the rule so we can use console directly for warnings and errors, but we still assume logs are dev artifacts that should not be checked in.